requirements.txt
requests
pandas
psycopg2-binary
streamlit
python-dotenv # To load .env file